ref: Cleanup tests for all actions
