+++ src/publish/__tests__/inputs.js
@@ -1,23 +1,7 @@
 /* eslint-env jest */
 
 const inputs = require("../inputs.js").default;
-
-function deepFreeze(object) {
-  // Retrieve the property names defined on object
-  const propNames = Object.getOwnPropertyNames(object);
-
-  // Freeze properties before freezing self
-
-  for (const name of propNames) {
-    const value = object[name];
-
-    if (value && typeof value === "object") {
-      deepFreeze(value);
-    }
-  }
-
-  return Object.freeze(object);
-}
+const deepFreeze = require("../deep-freeze.js").default;
 
 const inputsArgs = deepFreeze({
   context: {
